Create CompanyDetailsExtractor
## Company Details Extractor

The Company Details Extractor is a VBA code that allows you to extract information about any company, including its country of origin, CEO, founder, industry type, and more. This code is useful for anyone who needs to quickly gather information about a company for research or analysis purposes.

## Getting Started

To get started, simply download the CompanyDetailsExtractor.vba file and open it in Microsoft Excel. The code is written in VBA, so you will need to enable macros in order to run it.

## How to Use

Once you have opened the CompanyDetailsExtractor.vba file, follow these steps to extract information about a company:

1. In the "Company Details" worksheet, enter the name of the company you want to extract information about in cell B2.
2. Click the "Extract Details" button to run the code.
3. The code will extract information about the company from the internet and populate the corresponding cells in the worksheet.

## Features

The Company Details Extractor has the following features:

- Extracts information about any company, including its country of origin, CEO, founder, industry type, and more.
- Easy to use - simply enter the name of the company and click the "Extract Details" button.
- Works with any version of Microsoft Excel that supports VBA macros.